Common XRC: Modify parameters and composite devices (#806)
* Small adjustments to parameter model and config server
