+++ src/mx_bluesky/common/xrc_result.py
@@ -5,11 +5,27 @@ from collections.abc import Callable, Sequence
 from functools import partial
 
 import numpy as np
+from bluesky.callbacks import CallbackBase
+from event_model import RunStart
 
 from mx_bluesky.common.parameters.components import (
     MultiXtalSelection,
     TopNByMaxCountSelection,
 )
+
+
+class XRayCentreEventHandler(CallbackBase):
+    def __init__(self):
+        super().__init__()
+        self.xray_centre_results: Sequence[XRayCentreResult] | None = None
+
+    def start(self, doc: RunStart) -> RunStart | None:
+        if "xray_centre_results" in doc:
+            self.xray_centre_results = [
+                XRayCentreResult(**result_dict)
+                for result_dict in doc["xray_centre_results"]  # type: ignore
+            ]
+        return doc
 
 
 @dataclasses.dataclass
