+++ src/mx_bluesky/hyperion/experiment_plans/pin_tip_centring_plan.py
@@ -15,6 +15,7 @@ from dodal.devices.oav.utils import (
 )
 from dodal.devices.smargon import Smargon
 
+from mx_bluesky.common.utils.context import device_composite_from_context
 from mx_bluesky.common.utils.exceptions import SampleException
 from mx_bluesky.common.utils.log import LOGGER
 from mx_bluesky.hyperion.device_setup_plans.setup_oav import pre_centring_setup_oav
@@ -22,7 +23,6 @@ from mx_bluesky.hyperion.device_setup_plans.smargon import (
     move_smargon_warn_on_out_of_range,
 )
 from mx_bluesky.hyperion.parameters.constants import CONST
-from mx_bluesky.hyperion.utils.context import device_composite_from_context
 
 DEFAULT_STEP_SIZE = 0.5
